Button (#3)
Overview
Created working buttons, and began work on an array/for loop system to start and end each section, in terms of visibility.

HTML/CSS
index.css
Added Data-Type selector and visibility attributes.
index.html
Rearranged and added class bttn to sections, as to help forward the buttons.
Rearranged the jQuery and Bootstrap CDNs.

Rearranged the jQuery and Bootstrap CDNs in highscores.html

Javascript
index.js
Continued jQuery conversion.
Created an array of selectors with which to loop through.
Set some delay so the sections do not immediately disappear, for later warnings/notifications.
Verified that, for the time being, the first section is clickable and makes the next section visible the current one hidden.

Repository
n/a